WIP kernelci.config.load()
