+++ kci_test
@@ -23,7 +23,7 @@ import json
 import os
 import sys
 
-from kernelci.cli import Args, Command
+from kernelci.cli import Args, Command, parse_opts
 import kernelci
 import kernelci.config.lab
 import kernelci.config.test
@@ -258,6 +258,13 @@ class cmd_submit(Command):
 #
 
 if __name__ == '__main__':
+    opts = parse_opts("kci_test", "config/core", globals())
+    configs = kernelci.config.test.from_yaml(opts.yaml_config)
+    print(configs.keys())
+    exit(0)
+    status = opts.command(test_configs, lab_configs, opts)
+
+
     parser = argparse.ArgumentParser("kci_test")
     parser.add_argument("--yaml-test-configs",
                         default="config/core/test-configs.yaml",
