+++ kernelci/config/rootfs.py
@@ -18,7 +18,7 @@
 import sys
 import yaml
 
-from kernelci.config import YAMLObject
+from kernelci.config.base import YAMLObject
 
 
 class RootFS(YAMLObject):
@@ -136,10 +136,7 @@ class RootFSFactory(YAMLObject):
         return rootfs_cls.from_yaml(rootfs, kw)
 
 
-def from_yaml(yaml_path):
-    with open(yaml_path) as f:
-        data = yaml.safe_load(f)
-
+def from_yaml(data):
     rootfs_configs = {
         name: RootFSFactory.from_yaml(name, rootfs)
         for name, rootfs in data['rootfs_configs'].items()
